requirements.txt
fastapi==0.104.1
httptools==0.6.1
httpx==0.25.1
pip-chill==1.0.3
python-dotenv==1.0.0
pyyaml==6.0.1
rich==13.7.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
